Split entry_collections into two submodules
