+++ optimade/server/entry_collections/__init__.py
@@ -1,0 +1,4 @@
+from .entry_collections import EntryCollection
+from .mongo import MongoCollection, client, CI_FORCE_MONGO
+
+__all__ = ["EntryCollection", "MongoCollection", "client", "CI_FORCE_MONGO"]
